pyproject.toml
[tool.poetry]
name = "antlr-playground"
version = "0.1.0"
description = ""
authors = ["Oskar Korczak <oskar.korczak@protonmail.com>"]
readme = "README.md"
packages = [{include = "antlr_playground"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.25.2"
antlr4-python3-runtime = "^4.13.1"
pygls = "^1.0.2"
pytest = "^7.4.2"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
black = "^23.9.1"
mypy = "^1.5.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
